Re-visit test for `LibraryEntry`
